# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Jogo de Ações API
  version: "0.1.0"
  description: >
    API-first draft for Iteração 3 — routes derived from the four .feature files
    (create_competition, request_competition_entry, login, manage_competition_players).
    Written before any controller exists; adjust here first, then implement. Session-based
    auth (Spring Security, HttpSession/JSESSIONID cookie) established by consuming a login
    link — there is no password login.
servers:
  - url: /api
...
paths:
...
  /login-requests:
    post:
      tags: [login]
      operationId: requestLoginLink
      summary: Request a login link for an already-registered player or the administrator
      description: >
        Spec 05-005: `returnTo`, when present, is the page the caller was trying to reach
        before being asked to log in — consuming the resulting link redirects there instead
        of the role-based default. Must be a same-origin relative path (starts with a single
        `/`, never `//` right after); anything else is silently ignored, falling back to the
        default, since this endpoint requires no session and an unvalidated redirect target
        would be an open redirect.
      x-roles: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
                returnTo:
                  type: string
                  description: Same-origin relative path to redirect to after login — see the operation description.
      responses:
        '202': { description: Accepted — link sent by e-mail if the address is known }
  /login-links/{token}:
    get:
      tags: [login]
      operationId: consumeLoginLink
      summary: Validate and consume a login link
      description: >
        Establishes the HTTP session (Spring Security Authentication in SecurityContext)
        on success. Covers login.feature's device rules: a link only logs in the device
        that first used it (used on a different device → error, unless already logged in
        on the current device, which just redirects); a fresh link for the same player
        invalidates any previous unused link; exceeding the configured device limit ends
        the oldest session instead of blocking the new login.
      x-roles: []
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: >
            Valid link, already-registered player — session established, redirect target
            (a same-origin path) in the response body. See spec 05-005.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResult'
        '202':
          description: Valid link, new player — registration (name) still required
        '400':
          description: Link invalid or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Link already used to log in on a different device
  /login-links/{token}/registration:
    post:
      tags: [login]
      operationId: completeRegistration
      summary: Complete registration for a new player and confirm competition entry
      x-roles: []
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        '200':
          description: Registered, session established, added to the competition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResult'
        '400': { description: Link invalid or expired }
...
components:
  schemas:
...
    LoginResult:
      description: >
        Spec 05-005: `redirectTo` is a full same-origin path, not a symbol — either the
        caller's own `returnTo` (see requestLoginLink) when it was honored, or a role-based
        default (`/admin`, `/competitions/mine`, or `/competitions/{id}`) otherwise. No
        separate `competitionId` field — the id is already embedded in the path when
        applicable.
      type: object
      properties:
        redirectTo:
          type: string
    Error:
      type: object
      properties:
        message: { type: string }